Added functions [generateTaskId, createTaskCard, and handleDeleteTask]
Added functions the functions: generateTaskId, createTaskCard, and handleDeleteTask,

Task Id has iteration for when a card is added, removed and reset.
Removal iteration is handled in the delete card function.  Still needs logic for when the button is selected.

Create card uses provided content to add in the task information.