TA Help and Bug Fix
Worked with our TA Casey and the Class Instructor Dex to determine, locate and fix the bug I was struggling with.

Found that the todo and "to-do" identifiers were the cause of a display error with my task cards.

Fixed the assignment of the draggable class and applied it to the correct variable / event combos within my task id and create task functions.