Added dayjs, card color coding and css styling
I have added dayjs,to my createTaskCard function in order to setup card color coding and css styling.

Added a archive folder to save previous version in order to have an ability to revent my code if necessary.  This is something I have learned to do in the past assignment and is easier to use when resolving changes, version control and push updates to main.  In the future I will be setting up branches to help with this.